moved the meeting files into their correct folder
